Implented multithreading
Refactor some code and implemented multthreading on the computationaly
heavy portions of running a generation such as execution, mutation, and
crossover.
*Note: as the number of concurrent threads isn't bounded Java can
potentially take up 100% of the cpu utilization. If this occurs just
lower the priority of the Java Runtime enviorment to regain
responsivness on OS I/O.

Also added method of duplicate innovation tracking but even with
multithreading it is very time consuming when run on larger populations
and networks. Currently disabled.

Also added package which implements NEAT for image catogorization where
you place images in two folders and NEAT is trained to distinguish
between the two. For an example, this can be used to train a NN to
distinguish between appples and bannanas. *Note: I've gotten mixed
results so far but usually can achieve 80% accuracy with the current
implementation is the distinction is well defined.*
- This has nothing to do with Senior Design but is still interesting
enough to commit.