[Nov. 29, VEC]
GUI now shows a basic visualization of the network. For the sake of
testing, the for loop in the Main class is disabled. Edge weights not
taken into account. Nodes are color coded for clarity.

Upcoming tasks:
1)Clean up GUI. Have nodes on top of edges, and edges coming out from
the center of the node graphic. Issue where some nodes go out of screen.
Show edge weights via line thickness.
2)Currently, new GUI frame created per generation. Change to have 1 GUI
be updated to show progression of generations. 
3)Show additional information listed in timeline
4)Refine code. Add/ updated comments where necessary. Have G-classes be
based on general nodes and edges, rather than genes.